- changed blocks id's type (int -> byte) - added pre-compiler run task - first map parser tests :D

+++ pre-compiler/src/main/java/com/jvmfrog/precompiler/parser/MapSection.java
@@ -1,14 +1,14 @@
 package com.jvmfrog.precompiler.parser;
 
 public class MapSection extends Section {
-    public int[][] map;
+    public byte[][] map;
 
     @Override
     public void parse() {
-        map = new int[sectionData.size()][sectionData.get(0).data.length];
+        map = new byte[sectionData.size()][sectionData.get(0).data.length];
         for (int x = 0; x < sectionData.size(); x++) {
             for (int y = 0; y < sectionData.get(x).data.length; y++) {
-                map[x][y] = Integer.parseInt(sectionData.get(x).data[y]);
+                map[x][y] = Byte.parseByte(sectionData.get(x).data[y]);
             }
         }
     }
